Separate `FieldMerkleTree` with vectorized hashing
Moves the old `MerkleTree` into `generic_merkle_tree.rs`, and adds a separate `FieldMerkleTree` with field-specific logic, mainly vectorization based of `F::Packing`.
