+++ uni-stark/Cargo.toml
@@ -26,7 +26,7 @@ p3-baby-bear = { path = "../baby-bear" }
 p3-goldilocks = { path = "../goldilocks" }
 p3-keccak = { path = "../keccak" }
 p3-mersenne-31 = { path = "../mersenne-31" }
-p3-poseidon = { path = "../poseidon" }
+p3-poseidon2 = { path = "../poseidon2" }
 p3-symmetric = { path = "../symmetric" }
 rand = "0.8.5"
 tracing-subscriber = { version = "0.3.17", features = ["std", "env-filter"] }
